Create an abstract base api class
